backend/requirements.txt
fastapi
uvicorn
alembic
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
asyncpg
python-dotenv
loguru
pyyaml
typer
boto3
python-jose[cryptography]
PyJWT
requests
passlib[bcrypt]
python-multipart
pytest
pytest-asyncio
httpx
email-validator